[Feat] List페이지 컴포넌트로 변경
